refactor: extract shared events schemas to src/lib/events.ts
Consolidate EventTopicSchema, ListEventTopicsResultSchema, and
listEventTopics() into a shared module used by both scan.ts and the
event CLI commands, eliminating duplication from SMI-1540.
