+++ src/lib/bundle/scan.ts
@@ -68,39 +68,7 @@ function generateMockConfig(zodSchema: unknown): unknown {
 	return generateMockFromJsonSchema(jsonSchema)
 }
 
-const EventTopicSchema = z.object({
-	topic: z.string(),
-	name: z.string(),
-	description: z.string().optional(),
-	inputSchema: z.record(z.string(), z.unknown()).optional(),
-	eventSchema: z.record(z.string(), z.unknown()).optional(),
-})
-
-const ListEventTopicsResultSchema = z.object({
-	topics: z.array(EventTopicSchema),
-	nextCursor: z.string().optional(),
-})
-
-async function listEventTopics(client: Client) {
-	const capabilities = client.getServerCapabilities()
-	if (!capabilities?.experimental?.["ai.smithery/events"]) {
-		return { eventTopics: [] }
-	}
-
-	const eventTopics: z.infer<typeof EventTopicSchema>[] = []
-	let cursor: string | undefined
-	do {
-		// biome-ignore lint/suspicious/noExplicitAny: custom JSON-RPC method not in SDK types
-		const result = await (client.request as any)(
-			{ method: "ai.smithery/events/topics/list", params: { cursor } },
-			ListEventTopicsResultSchema,
-		)
-		eventTopics.push(...result.topics)
-		cursor = result.nextCursor
-	} while (cursor)
-
-	return { eventTopics }
-}
+import { listEventTopics } from "../events"
 
 export interface ScanResult {
 	serverCard?: ServerCard
